L4_INTEL_80216_INDICATION typedef redefinition
The variable holding L3_L4_OPCODE_* value undergoes various typecasting
changes as it moves along the call stack.  The end result, on big-endian
machines, is that the variable holding the opcode (indication_id) will
always equal 0.  Changing the typedef from UINT16 to UINT32 resolves
this problem.

+++ InfraStack/OSAgnostic/Product/AppSrvInfra/L4MsgProxy.h
@@ -57,7 +57,7 @@ extern L5_CONNECTION l4Proxy_Connection;
 extern tL5DispatcherFunctions *l4Proxy_pFuncs;
 
 
-typedef UINT16 L4_INTEL_80216_INDICATION;
+typedef UINT32 L4_INTEL_80216_INDICATION;
 
 
 // AppSrv Agent Interface:
